clue card chagne

+++ module/clue-cards.ts
@@ -1,8 +1,11 @@
-import { CityDB } from "./city-db.mjs";
-import {CityLogger} from "./city-logger.mjs";
+import { CityHelpers } from "./city-helpers.js";
+import { CityActor } from "./city-actor.js";
+import { CityDB } from "./city-db.js";
+import {CityLogger} from "./city-logger.js";
 import {CitySettings} from "./settings.js";
 
 export class ClueChatCards {
+	static clickLock : boolean = false;
 
 	/**  posts a clue card to the chat
 	@param {Object} templateData
@@ -11,7 +14,7 @@ export class ClueChatCards {
 	@param {string} templateData.method The method of the clue
 	@param {string} templateData.source The source for the clue
 	*/
-	static async postClue (templateData) {
+	static async postClue (templateData: Record<string, unknown> & {actorId: string}) {
 		const actor = CityDB.getActorById(templateData.actorId);
 		// const	templateData= {actorId, metaSource, method};
 		if (!templateData.metaSource)
@@ -19,12 +22,12 @@ export class ClueChatCards {
 		if (!actor)
 			throw new Error(`Couldnt find actor Id ${templateData?.actorId}`);
 		const html = await renderTemplate("systems/city-of-mist/templates/parts/clue-reveal.hbs", templateData);
-		await CityLogger.sendToChat2(html, {actor});
+		await CityLogger.sendToChat2(html, {actor: actor.id});
 	}
 
 	static CLUE_LOCK = false;
 
-	static async updateClue (html, newdata ={}) {
+	static async updateClue (html: JQuery<HTMLElement>, newdata ={}) {
 		if (ClueChatCards.CLUE_LOCK) return;
 		ClueChatCards.CLUE_LOCK = true;
 		const messageId = html.data("messageId");
@@ -39,23 +42,25 @@ export class ClueChatCards {
 		if (!metaSource)
 			console.warn("No metasource for clue");
 		const new_html = await renderTemplate("systems/city-of-mist/templates/parts/clue-reveal.hbs", templateData);
-		const user = message.user;
+		const user = message?.user;
 		try {
+			if (!message) {
+				throw new Error("No message");
+			}
 			await message.delete();
-			const actor = CityDB.getActorById(actorId);
+			const actor = CityDB.getActorById(actorId) as CityActor;
 			let whisperTarget = "";
 			if (CitySettings.whisperClues()) {
 				whisperTarget = message.user.id;
 			}
-			const msg = await CityLogger.sendToChat2(new_html, {actor, token: "", alias:actor ? actor.getDisplayedName() : "clue", altUser: user}, whisperTarget);
-			// msg.user = user;
+			await CityLogger.sendToChat2(new_html, {actor: actor.id, token: "", alias:actor ? actor.getDisplayedName() : "clue", altUser: user}, whisperTarget);
 		} catch (err) {
 			console.error(err);
 		}
 		ClueChatCards.CLUE_LOCK = false;
 	}
 
-	static async clueEditButtonHandlers(_app, html, _data) {
+	static async clueEditButtonHandlers(_app: unknown, html: JQuery<HTMLElement>, _data: unknown) {
 		if ($(html).find(".clue-reveal").length == 0)
 			return true;
 		$(html).find(".question-part .submit-button").click (
@@ -100,45 +105,48 @@ export class ClueChatCards {
 		return true;
 	}
 
-	static async clue_bankClue(html) {
+	static async clue_bankClue(html: JQuery<HTMLElement>) {
 		const messageId = html.data("messageId");
 		const actorId = $(html).find(".clue-reveal").data("actorId");
-		const huge_method = $(html).find(".clue-reveal").data("method");
+		const huge_method = $(html).find(".clue-reveal").data("method") as String;
 		const source = $(html).find(".clue-reveal").data("source");
 		const partial_clue = $(html).find(".clue-reveal").data("partialClue");
 		const metaSource = $(html).find(".clue-reveal").data("metaSource");
-		const actor = CityDB.findById(actorId, "Actor");
+		const actor = CityDB.findById(actorId, "Actor") as CityActor;
 		if (!actor ) throw new Error(`Couldn't find Actor ${actorId}`);
 		const [name, method]   = huge_method.split(":").map (x=> x.trim());
 		const clueData = {partial: partial_clue, name, method, source};
 		const message = game.messages.get(messageId);
+		if (!message) {
+			throw new Error("Can't find message");
+		}
 		const new_html = await renderTemplate("systems/city-of-mist/templates/parts/clue-reveal.hbs", {banked:true});
 		const msg = await  message.update( {content: new_html});
 		await ui.chat.updateMessage( msg, false);
 		await actor.createClue(metaSource, clueData);
 	}
 
-	static async clue_submitAnswer(html) {
+	static async clue_submitAnswer(html: JQuery<HTMLElement>) {
 		const answer = $(html).find(".answer-part .answer-input").val();
 		await this.updateClue( html, {answer});
 	}
 
-	static async clue_editAnswer(html) {
+	static async clue_editAnswer(html: JQuery<HTMLElement>) {
 		const question = $(html).find(".clue-reveal").data("question");
 		const answer = $(html).find(".clue-reveal").data("answer");
 		await this.updateClue( html, {question, partial_answer_text: answer});
 	}
 
-	static async clue_addToJournal(html) {
+	static async clue_addToJournal(html: JQuery<HTMLElement> ) {
 		const question = $(html).find(".clue-reveal").data("question");
 		const answer = $(html).find(".clue-reveal").data("answer");
 		const actorId = $(html).find(".clue-reveal").data("actorId");
-		const actor = CityDB.findById(actorId, "Actor");
+		const actor = CityDB.findById(actorId, "Actor") as CityActor;
 		if (await actor.addClueJournal(question, answer))
 			await CityHelpers.playWriteJournal();
 	}
 
-	static async clue_partialClueCheckbox(html) {
+	static async clue_partialClueCheckbox(html: JQuery<HTMLElement>) {
 		if (!game.user.isGM) return;
 		const partial_clue = $(html).find(".partial-clue").is(":checked");
 		const answer = $(html).find(".clue-reveal").data("answer");
@@ -153,13 +161,13 @@ export class ClueChatCards {
 		this.clickLock = false;
 	}
 
-	static async clue_refundClue(html) {
+	static async clue_refundClue(html: JQuery<HTMLElement>) {
 		const question_rejected= true;
 		const question = "";
 		await this.updateClue( html, {question, question_rejected} );
 	}
 
-	static async clue_submitQuestion(html) {
+	static async clue_submitQuestion(html: JQuery<HTMLElement>) {
 		const question = $(html).find(".question-part .question-input").val();
 		await this.updateClue( html, {question});
 	}
